+++ source/exceeds_expectations/pretty_print.d
@@ -11,7 +11,7 @@ import std.traits;
 
 /// Prints the given value in a nice, human readable format. If receiving a
 /// string, it will output the string with bold double quotes indicating the
-/// start and end of the string. if the returned string had line breaks, it is
+/// start and end of the string. If the returned string had line breaks, it is
 /// guaranteed to also start and end with a line break.
 package string prettyPrint(T)(T value)
 out(result; !(result.canFind('\n')) || (result.endsWith("\n") && result.startsWith("\n")))