+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] · Not Yet Released
+
+### Changed
+- `toBe` shows prints difference between received and expected values (like toEqual) if it received non-reference types.
+- Structs are now pretty-printed using the same format as class objects.
+- The pretty-print format for classes and structs now includes the keyword `class` or `struct`.
+- Fixed highlighted output appearing unreadable in some color schemes.
+
+
 ## [0.6.0] · 2021-05-23
 
 ### Added
@@ -108,3 +117,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.4.2]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.4.2
 [0.5.0]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.5.0
 [0.6.0]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.6.0
+[0.7.0]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.7.0